Adds a new report to the list of reports available
For the prototype, this commit contains everything needed to generate a
report for the quality of the metadata.  Currently it checks a few
fields and assigns the dataset a score, averaging them for the
organisation and showing 5 of the worst so that they can be fixed (at
source, not in our catalogue).