Improve quality report
The quality report now implements some of the EA Data Quality Action
Plan, where the checks can be made programmatically. Not all of the
checks have yet been implemented, but there are enough to show the
intended use for that report.

Also, a fly-by fix to hide 'private-resources' and 'gdpr' and
'reference' extras from showing in the Additional info section.